refactor: update import paths for I18nextScannerService
Change the import paths for I18nextScannerService to reflect new
directory structure, improving codebase organization.
